make dump-object go gettable as ebpf-dump
